Add files via upload
This is a Dynamic DNS you can run locally if you use DreamHost as your web host. Run it locally on a linux server which has access to the internet (no ports needed to be opened specifically just http(s). Create a file called domains.txt in the same folder as this script. In domains.txt enter the fqdn you want to update. IE: asdf.example.com or just example.com This will use the API from DreamHost to update the IP's by checking what the old IP was and if its different updating to the new IP. Set this as a cronjob and run as often as you'd like to keep DNS up to date. As this isn't changing records all the time a run rate of ~1 minute for small list of domains (<10) is likely fine. If large list you'll have to let it run longer in case of network congestion (API side or locally).